enhance security by checking user id with album owner id

+++ app/Http/Controllers/V1/AlbumController.php
@@ -7,15 +7,16 @@ use App\Models\Album;
 use App\Http\Requests\StoreAlbumRequest;
 use App\Http\Requests\UpdateAlbumRequest;
 use App\Http\Resources\V1\AlbumResource;
+use Request;
 
 class AlbumController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        return AlbumResource::collection(Album::all());
+        return AlbumResource::collection(Album::where()->paginate());
     }
 
     /**
@@ -30,9 +31,12 @@ class AlbumController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Album $album)
+    public function show(Request $request, Album $album)
     {
-        return $album;
+        if($request->user()->id != $album->user_id) {
+            return abort(403,'Unautorised');
+        }
+        return new AlbumResource($album);
     }
 
     /**
@@ -40,6 +44,9 @@ class AlbumController extends Controller
      */
     public function update(UpdateAlbumRequest $request, Album $album)
     {
+        if($request->user()->id != $album->user_id) {
+            return abort(403,'Unautorised');
+        }
         $album->update($request->all());
 
         return new AlbumResource($album);
@@ -48,12 +55,15 @@ class AlbumController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Album $album)
+    public function destroy(Request $request,Album $album)
     {
+        if($request->user()->id != $album->user_id) {
+            return abort(403,'Unautorised');
+        }
         $album->delete();
 
         return response('',204);
     }
 
-    
+
 }
